tests/govtool-backend/requirements.txt
pytest-github-report==0.0.1
python-dotenv==1.0.0
requests==2.31.0
allure-pytest==2.13.5